requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
audioop-lts==0.2.1
colorlog==6.9.0
confspec==0.0.3
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.6.0
idna==3.10
linkd==0.0.7
multidict==6.4.4
propcache==0.3.1
python-dotenv==1.1.0
yarl==1.20.0
